# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для платформы изучения английского языка. Используется Dictionary API для получения данных о словах.",
        "title": "English Learning Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/vocabulary": {
            "get": {
                "description": "Возвращает список всех слов в личном словаре или слова для конкретного пользователя (если указан user_id)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "personal-vocabulary"
                ],
                "summary": "Получить все слова личного словаря",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID) для фильтрации слов",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PersonalVocabulary"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новое слово в личном словаре пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "personal-vocabulary"
                ],
                "summary": "Создать слово в личном словаре",
                "parameters": [
                    {
                        "description": "Данные для создания слова",
                        "name": "vocabulary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreatePersonalVocabularyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.PersonalVocabulary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/vocabulary/review": {
            "get": {
                "description": "Возвращает слова, которые нужно повторить (next_review \u003c= now)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "personal-vocabulary"
                ],
                "summary": "Получить слова для повторения",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PersonalVocabulary"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/vocabulary/{id}": {
            "get": {
                "description": "Возвращает информацию о слове в личном словаре по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "personal-vocabulary"
                ],
                "summary": "Получить слово по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vocabulary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PersonalVocabulary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет информацию о слове в личном словаре",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "personal-vocabulary"
                ],
                "summary": "Обновить слово в личном словаре",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vocabulary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления слова",
                        "name": "vocabulary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdatePersonalVocabularyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PersonalVocabulary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет слово из личного словаря пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "personal-vocabulary"
                ],
                "summary": "Удалить слово из личного словаря",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vocabulary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/vocabulary/{id}/stats": {
            "put": {
                "description": "Обновляет счетчики правильных/неправильных ответов для слова",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "personal-vocabulary"
                ],
                "summary": "Обновить статистику слова",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vocabulary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления статистики",
                        "name": "stats",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateStatsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PersonalVocabulary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.CreatePersonalVocabularyRequest": {
            "type": "object",
            "required": [
                "translation",
                "user_id",
                "word"
            ],
            "properties": {
                "audio_url": {
                    "type": "string"
                },
                "example": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "phonetic": {
                    "type": "string"
                },
                "status": {
                    "description": "'new', 'learning', 'learned'",
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "translation": {
                    "type": "string"
                },
                "user_id": {
                    "description": "UUID в виде строки",
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdatePersonalVocabularyRequest": {
            "type": "object",
            "required": [
                "translation",
                "word"
            ],
            "properties": {
                "audio_url": {
                    "type": "string"
                },
                "example": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "phonetic": {
                    "type": "string"
                },
                "status": {
                    "description": "'new', 'learning', 'learned'",
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "translation": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        },
        "handlers.UpdateStatsRequest": {
            "type": "object",
            "required": [
                "is_correct"
            ],
            "properties": {
                "is_correct": {
                    "type": "boolean"
                }
            }
        },
...
        "models.PersonalVocabulary": {
            "type": "object",
            "properties": {
                "audio_url": {
                    "type": "string"
                },
                "correct_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "example": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_seen": {
                    "type": "string"
                },
                "next_review": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "phonetic": {
                    "type": "string"
                },
                "status": {
                    "description": "'new', 'learning', 'learned'",
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "translation": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                },
                "wrong_count": {
                    "type": "integer"
                }
            }
        },
...
    }
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Type Bearer followed by a space and JWT token."
        }
    }
}